refactor: Security AuthenticationProvider 흐름 코드 구조 수정
